adding exponential
